+++ src/components/Buckets/index.js
@@ -5,7 +5,6 @@ import './index.css';
 
 class Buckets extends Component {
   render() {
-    console.log(this.props);
     if (this.props.isAuth !== true) {
       return (<Redirect to="/login" />);
     }
